requirements.txt
python-telegram-bot[job-queue]>=20.0
openai>=1.0.0
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
